+++ src/main/groovy/fr/nereide/completion/provider/xml/MenuNameCompletionProvider.groovy
@@ -14,7 +14,7 @@ import com.intellij.psi.xml.XmlTag
 import com.intellij.util.ProcessingContext
 import fr.nereide.dom.element.menu.Menu
 import fr.nereide.dom.file.MenuFile
-import fr.nereide.project.ProjectServiceInterface
+import fr.nereide.project.OfbizProjectHelper
 import fr.nereide.project.utils.MiscUtils
 import org.jetbrains.annotations.NotNull
 
@@ -23,7 +23,7 @@ class MenuNameCompletionProvider extends CompletionProvider<CompletionParameters
 
     @Override
     protected void addCompletions(@NotNull CompletionParameters parameters, @NotNull ProcessingContext context, @NotNull CompletionResultSet result) {
-        ProjectServiceInterface ps = parameters.getPosition().getProject().getService(ProjectServiceInterface.class)
+        OfbizProjectHelper ph = OfbizProjectHelper.getInstance(parameters.position.project)
         PsiElement myElement = parameters.getPosition()
         List<Menu> menus
         XmlElement myAttrValue
@@ -35,7 +35,7 @@ class MenuNameCompletionProvider extends CompletionProvider<CompletionParameters
         XmlTag parentTag = PsiTreeUtil.getParentOfType(myAttrValue, XmlTag.class)
         if (parentTag.getAttribute('location')) {
             XmlAttributeValue menuLocationAttr = parentTag.getAttribute('location').getValueElement()
-            menus = ps.getDomElementListFromFileAtLocation(menuLocationAttr.value, MenuFile.class)
+            menus = ph.getDomElementListFromFileAtLocation(menuLocationAttr.value, MenuFile.class)
         } else {
             menus = []
         }