+++ src/main/groovy/fr/nereide/reference/xml/GenericXmlReference.groovy
@@ -21,23 +21,23 @@ import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiReferenceBase
 import com.intellij.psi.xml.XmlAttributeValue
 import com.intellij.util.xml.DomManager
-import fr.nereide.project.ProjectServiceInterface
+import fr.nereide.project.OfbizProjectHelper
 
 abstract class GenericXmlReference extends PsiReferenceBase<XmlAttributeValue> {
 
     DomManager dm
-    ProjectServiceInterface ps
+    OfbizProjectHelper ph
 
     GenericXmlReference(XmlAttributeValue element, TextRange textRange, boolean soft) {
         super(element, textRange, soft)
         dm = DomManager.getDomManager(element.getProject())
-        ps = element.getProject().getService(ProjectServiceInterface.class)
+        ph = OfbizProjectHelper.getInstance(element.project)
     }
 
     GenericXmlReference(XmlAttributeValue element, boolean soft) {
         super(element, soft)
         dm = DomManager.getDomManager(element.getProject())
-        ps = element.getProject().getService(ProjectServiceInterface.class)
+        ph = OfbizProjectHelper.getInstance(element.project)
     }
 
 }