+++ src/main/groovy/fr/nereide/reference/xml/provider/GroovyServiceMethodReferenceProvider.groovy
@@ -5,7 +5,7 @@ import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiReference
 import com.intellij.psi.xml.XmlAttributeValue
 import com.intellij.util.ProcessingContext
-import fr.nereide.project.ProjectServiceInterface
+import fr.nereide.project.OfbizProjectHelper
 import fr.nereide.reference.xml.GroovyServiceDefReference
 import org.jetbrains.annotations.NotNull
 import org.jetbrains.plugins.groovy.lang.psi.GroovyFile
@@ -14,10 +14,10 @@ class GroovyServiceMethodReferenceProvider extends JavaMethodReferenceProvider {
 
     @Override
     PsiReference[] getReferencesByElement(@NotNull PsiElement element, @NotNull ProcessingContext context) {
-        ProjectServiceInterface ps = element.getProject().getService(ProjectServiceInterface.class)
+        OfbizProjectHelper ph = OfbizProjectHelper.getInstance(element.project)
         String locationAttr = getClassLocation(element)
         if (!locationAttr) return PsiReference.EMPTY_ARRAY
-        PsiFile targetedFile = ps.getPsiFileAtLocation(locationAttr)
+        PsiFile targetedFile = ph.getPsiFileAtLocation(locationAttr)
         if (!targetedFile || !(targetedFile instanceof GroovyFile)) return PsiReference.EMPTY_ARRAY
         return new GroovyServiceDefReference(element as XmlAttributeValue,
                 (targetedFile as GroovyFile).getScriptClass(),